[Hotfix-14-DTO-계층-변경-Little-SNS-MS] 테스트 코드
