+++ little-SNS-MS/src/test/java/com/littleSNSMS/infra/db/entities/PostEntityTest.java
@@ -1,7 +1,8 @@
 package com.littleSNSMS.infra.db.entities;
 
 import com.littleSNSMS.domain.Post;
-import com.littleSNSMS.domain.dto.PostDTO;
+
+import com.littleSNSMS.domain.service.PostService;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -11,7 +12,7 @@ class PostEntityTest {
     @Test
     void post() {
         // Given
-        Post post = Post.post(PostDTO.Create.of("Hello world!", "ID", "test@test.com"));
+        Post post = Post.post(PostService.Create.of("Hello world!", "ID", "test@test.com"));
 
         // When
         PostEntity result = PostEntity.post(post);
